requirements.txt
streamlit>=1.25.0
google-generativeai>=0.7.2
python-dotenv>=1.0.0
sentry-sdk>=1.10.0
